foundry vtt regel exporter: - renames wrapper file
